remove models from bluesky
