+++ config/settings/common.py
@@ -110,8 +110,7 @@ SEED_CORE_APPS = (
     'seed.lib.superperms.orgs',
     'seed.audit_logs',
     'seed.cleansing',
-    'seed.functional',
-    'seed.bluesky'
+    'seed.functional'
 )
 
 # Apps with tables created by migrations, but which 3rd-party apps depend on.
