+++ seed/models/__init__.py
@@ -1,0 +1,7 @@
+# Import all the models in this folder
+
+# TODO: Should we move all the views to the, well, views folder?
+from .models import *
+from .cycle import *
+from .property import *
+from .tax_lot import *
